requirements.txt
asgiref==3.3.4
Django==3.2.3
django-cpf==0.1.0
django-crispy-forms==1.11.2
django-localflavor==3.0.1
django-phone-field==1.8.1
python-stdnum==1.16
pytz==2021.1
sqlparse==0.4.1
